Phase 4 & 5: extract media file-type, archive, and deepfake helpers into submodules
